refactor: Flash plugins - II
leave just a shim behind
