+++ mpflash/flash/stm32.py
@@ -1,24 +1,4 @@
-"""Flash STM32 boards using pydfu.
+"""Compatibility shim — the implementation moved to :mod:`mpflash.flash.builtins.dfu`."""
 
-This helper is now the internal implementation of
-:class:`mpflash.flash.builtins.dfu_backend.DFUBackend`. New code should call
-:func:`mpflash.flash.flash_mcu` with ``method=FlashMethod.DFU``.
-"""
-
-from pathlib import Path
-
-from loguru import logger as log
-
-from mpflash.mpremoteboard import MPRemoteBoard
-
-# from .flash_stm32_cube import flash_stm32_cubecli
-from .stm32_dfu import dfu_init, flash_stm32_dfu
-
-
-def flash_stm32(mcu: MPRemoteBoard, fw_file: Path, *, erase: bool):
-    # sourcery skip: lift-return-into-if
-    dfu_init()
-    if updated := flash_stm32_dfu(mcu, fw_file=fw_file, erase=erase):
-        mcu.wait_for_restart()
-        log.success(f"Flashed {mcu.board_id} {mcu.version} to {mcu.serialport}.")
-    return updated
+from mpflash.flash.builtins.dfu import *  # noqa: F401,F403
+from mpflash.flash.builtins.dfu import flash_stm32  # noqa: F401
